week5assignment/requirements.txt
pyspark
psycopg2-binary
pandas
sqlalchemy
azure-storage-blob
